requirements.txt
click==8.1.8
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.3.5
rich==14.0.0
shellingham==1.5.4
smmap==5.0.2
SQLAlchemy==2.0.41
typer==0.15.4
typing_extensions==4.13.2
